SE0_Analyzer/requirements.txt
# Flask Web Framework
flask==3.0.0
Werkzeug==3.0.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/LLM Integration
google-generativeai==0.3.1

# Additional utilities
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2